fix(x64): apply proper warning fixes across server game/db/libsql source
Parallel pass fixing 64-bit truncation/conversion warnings across 107
server source files (886 fixes): explicit casts for bounded
size_t/ptrdiff_t index and length values, pointer<->integer via
intptr_t/uintptr_t, deprecated CRT/Winsock migrations, and format
specifier corrections. Plus stdafx.h maps strcasecmp/strncasecmp to
_stricmp/_strnicmp (Windows-guarded, <string.h> included) which clears
C4996 at every str_cmp/TOKEN call site.

Server x64: 0 compile errors; warnings 2468 -> ~320. Remaining are
time_t->32-bit timestamp and GoldType->32-bit money narrowings that need
coordinated type widening, plus the liblua .c bounded casts -- handled
next.

+++ Lead-Server-Source/db/src/Cache.cpp
@@ -224,12 +224,12 @@ void CItemPriceListTableCache::UpdateList(const TItemPriceListTable* pUpdateList
 			sizeAddOldDataSize = tmpvec.size();
 
 		thecore_memcpy(m_data.aPriceInfo + pUpdateList->byCount, &tmpvec[0], sizeof(TItemPriceInfo) * sizeAddOldDataSize);
-		m_data.byCount += sizeAddOldDataSize;
+		m_data.byCount += static_cast<BYTE>(sizeAddOldDataSize);
 
-		nDeletedNum = tmpvec.size() - sizeAddOldDataSize;
+		nDeletedNum = static_cast<int>(tmpvec.size() - sizeAddOldDataSize);
 	}
 	else
-		nDeletedNum = tmpvec.size();
+		nDeletedNum = static_cast<int>(tmpvec.size());
 
 	m_bNeedQuery = true;
 
@@ -256,8 +256,8 @@ void CItemPriceListTableCache::OnFlush()
 	for (int idx = 0; idx < m_data.byCount; ++idx)
 	{
 		snprintf(szQuery, sizeof(szQuery),
-				"INSERT INTO myshop_pricelist%s(owner_id, item_vnum, price) VALUES(%u, %u, %u)", 
-				GetTablePostfix(), m_data.dwOwnerID, m_data.aPriceInfo[idx].dwVnum, m_data.aPriceInfo[idx].dwPrice);
+				"INSERT INTO myshop_pricelist%s(owner_id, item_vnum, price) VALUES(%u, %u, %lld)",
+				GetTablePostfix(), m_data.dwOwnerID, m_data.aPriceInfo[idx].dwVnum, static_cast<long long>(m_data.aPriceInfo[idx].dwPrice));
 		CDBManager::instance().ReturnQuery(szQuery, QID_ITEMPRICE_SAVE, 0, NULL);
 	}
