+++ Lead-Server-Source/game/src/MarkManager.cpp
@@ -188,12 +188,12 @@ DWORD CGuildMarkManager::__AllocMarkID(DWORD guildID)
 
 DWORD CGuildMarkManager::GetMarkImageCount() const
 {
-	return m_mapIdx_Image.size();
+	return static_cast<DWORD>(m_mapIdx_Image.size());
 }
 
 DWORD CGuildMarkManager::GetMarkCount() const
 {
-	return m_mapGID_MarkID.size();
+	return static_cast<DWORD>(m_mapGID_MarkID.size());
 }
 
 // SERVER
@@ -203,8 +203,8 @@ void CGuildMarkManager::CopyMarkIdx(char * pcBuf) const
 
 	for (std::map<DWORD, DWORD>::const_iterator it = m_mapGID_MarkID.begin(); it != m_mapGID_MarkID.end(); ++it)
 	{
-		*(pwBuf++) = it->first; // guild id
-		*(pwBuf++) = it->second; // mark id
+		*(pwBuf++) = static_cast<WORD>(it->first); // guild id
+		*(pwBuf++) = static_cast<WORD>(it->second); // mark id
 	}
 }
 
@@ -358,13 +358,13 @@ void CGuildMarkManager::SaveSymbol(const char* filename)
 		return;
 	}
 
-	DWORD symbolCount = m_mapSymbol.size();
+	DWORD symbolCount = static_cast<DWORD>(m_mapSymbol.size());
 	fwrite(&symbolCount, 4, 1, fp);
 
 	for (std::map<DWORD, TGuildSymbol>::iterator it = m_mapSymbol.begin(); it != m_mapSymbol.end(); ++it)
 	{
 		DWORD guildID = it->first;
-		DWORD dwSize = it->second.raw.size();
+		DWORD dwSize = static_cast<DWORD>(it->second.raw.size());
 		fwrite(&guildID, 4, 1, fp);
 		fwrite(&dwSize, 4, 1, fp);
 		fwrite(&it->second.raw[0], 1, dwSize, fp);
